Slim VM session and stdlib modules
